Move buffer loading to a separate pair of source files.
