+++ src/app/gltf_viewer/gltf_viewer/GltfRendering.cpp
@@ -1,11 +1,10 @@
 #include "GltfRendering.h"
 
+#include "LoadBuffer.h"
 #include "Logging.h"
 #include "Shaders.h"
 
 #include <renderer/Uniforms.h>
-
-#include <handy/Base64.h>
 
 #include <math/Transformations.h>
 
@@ -55,67 +54,7 @@ const std::map<gltf::Accessor::ElementType, VertexAttributeLayout> gElementTypeT
 // Helper functions
 //
 
-std::vector<std::byte> loadInputStream(std::istream && aInput, std::size_t aByteLength, const std::string & aStreamId)
-{
-    constexpr std::size_t gChunkSize = 128 * 1024;
-
-    std::vector<std::byte> result(aByteLength);
-    std::size_t remainingBytes = aByteLength;
-    
-    while (remainingBytes)
-    {
-        std::size_t readSize = std::min(remainingBytes, gChunkSize);
-        if (!aInput.read(reinterpret_cast<char *>(result.data()), readSize).good())
-        {
-            throw std::runtime_error{"Problem reading '" + aStreamId + "': "
-                // If stream is not good(), yet converts to 'true', it means only eofbit is set.
-                + (aInput ? "stream truncacted." : "read error.")};
-        }
-        remainingBytes -= readSize;
-    }
-    return result;
-}
-
-
-std::vector<std::byte> loadBufferData(Const_Owned<gltf::BufferView> aBufferView)
-{
-    auto buffer = aBufferView.get(&gltf::BufferView::buffer);
-
-    if (!buffer->uri)
-    {
-        ADLOG(gPrepareLogger, critical)
-             ("Buffer #{} does not have target defined.", aBufferView->buffer);
-        throw std::logic_error{"Buffer was expected to have an Uri."};
-    }
-
-    constexpr const char * base64Prefix{"base64,"};
-
-    gltf::Uri uri = *buffer->uri;
-    switch(uri.type)
-    {
-    case gltf::Uri::Type::Data:
-    {
-        ADLOG(gPrepareLogger, trace)("Buffer #{} data is read from a data URI.", aBufferView->buffer);
-        std::string_view encoded{uri.string};
-        encoded.remove_prefix(encoded.find(base64Prefix) + std::strlen(base64Prefix));
-        return handy::base64::decode(encoded);
-    }
-    case gltf::Uri::Type::File:
-    {
-        ADLOG(gPrepareLogger, trace)("Buffer #{} data is read from file {}.", aBufferView->buffer, uri.string);
-        return loadInputStream(
-            std::ifstream{buffer.getFilePath(&gltf::Buffer::uri).string(),
-                          std::ios_base::in | std::ios_base::binary},
-            buffer->byteLength, 
-            uri.string);
-    }
-    default:
-        throw std::logic_error{"Invalid uri type."};
-    }
-}
-
-
-// TODO accept the buffer view directly, but it require to embed the index in it for debug print.
+// TODO accept the buffer view directly, but it first require to embed the index in it for debug print.
 template <class T_buffer>
 std::pair<Const_Owned<gltf::BufferView>, T_buffer>
 prepareBuffer_impl(Const_Owned<gltf::Accessor> aAccessor)
